Changed each instance of
  ca_pend_io( time );

to

  ca_pend_io( time );
  ca_pend_event( 0.00001 );     (ok I know the resolution is unreasonable,
                                 but I'm covered for a long time)


Without the pend event, we observed the following weird behavior:

The first activation of a display was quick. After deactivating the
display and activating any other one (even the same one), from that
moment on, the new activation would be slow and events would kind
of dribble in.

With the pend event, this behavior went away (at least so far).