Change version to 1-7-0d
When objects are stacked on top of one another, it can be difficult to
select a particular object. Previously, it was necessary to know
beforehand the stacking order. To select the Nth object, you would
left-click to select one object (the top one) and then control-left-click
N-1 times to cycle through the additional objects. Now, when you
control-left-click the object name is displayed below and to the right
of the selected object. It is now therefore much easier to find the object of
interest in a bunch of objects stacked together.