Added another complexity to symbols:
If "Binary Truth Table" is not checked and multiple PVs are specified then
an AND, an XOR, and a shift operation is applied to each PV value as follows

  val = 0;

  if ( "and mask" != 0 ) val = val & "and mask"

  if ( "xor mask" != 0 ) val = val ^ "xor mask"

  if ( "shift count" != 0 ) val = val << "shift count"

  ( negative shift count implies >> )

The final value is the logical OR of all values as computed above.

=================================

If "Binary Truth Table" is checked then behavior is unchanged, i.e.
first PV influences bit 0 of the final value such that bit = 0 if
PV value == 0; bit = 1 if PV value != 0;

Second PV influences bit 1 and so on.

=================================

If only 1 PV is specified and the "and mask" is 0, the "xor mask" is 0,
and the "shift count" is 0, then the final value is a double given as

  final value = (double) "value of PV"