Version 1-9-5g
app_pkg.cc:

When ctl pv is set to a screen name, and the screen is
already opened, the window is raised and the xwindow id
preceeded by a space is written back to the ctl pv.

To support the above, if the first character of the ctl pv
is a space, edm ignores it and does not attempt to use the
contents to open a display file.