Version 1-9-5w
group.cc: modified updateGroup

The following bug was fixed:

If a group of objects contains a symbol or dynamic symbol and the group
is copied from one window to another, the window object does not get
updated for the symbol or dynamic symbol.

This can cause edm to crash.

Note that if edm is running and no edits have been performed, there is
no problem.