Version 1-9-5zb
When an edm screen contains references to fonts that do not
exist, a simple font substitution is performed. This results
in the screen "changed" flag being set true.

When the screen is closed, the "save changes?" dialog is thus invoked.

This makes no sense when edm is executed in "noedit" mode.

act_win.cc was modified so that the "changed" flag is never set
true when edm is executed in "noedit" mode.