Version 1-9-5zd
Added font groups to the version 3 font file

You may now specify fonts like this:

helvetica={
-urw-helvetica-(medium,bold)-(o,r)-normal--0-(80,90,100)-75-75-p-0-*-*
-*-helvetica-(medium,bold)-(o,r)-normal--0-(80,90,100)-75-75-p-0-*-*
}

In this example, if the urw helvetica font is present then it is used
otherwise any helvetica font is used.

Edm attempts to match all font specs exactly and uses the first one
that succeeds. If all specs fail to match, edm uses the last line in
the group and trys to find a fallback set of fonts that may even contain
members of another family (this is done with single font specs also).