Version 1-9-6d
Edm will now generate comments from a comment template file.

For example:

Suppose the file "edmComments" contains:

  # $Revision$
  #

Then, you may set the environment variable EMDCOMMENTS to /somepath/edmComments
and edm will add the following to the beginning of an edl file on save:

  # <<<edm-generated-comments>>>
  #
  # $Revision$
  #

Subsequently, when edm sees the "<<<edm-generated-comments>>>"
signature in the comment area, additional comments will not
be generated.