Version 1-10-0q-beta
Fixes:

The slider and motifslider widgets were not updating properly when an
external agent was rapidly changing the value of the control pv.

Line widget contained uninitialized variables.

New Feature:

Files may be retrieved from http server.

To turn this on, add -DUSECURL=1 to the compiler flags
and link with libcurl.so (-lcurl).

At runtime, set the environment variable
EDMHTTPDOCROOT to a list of urls sepatated by
the "|" character.