Version 1-10-0y
app_pkg.cc, related_display.cc, pv_inspector.cc - Add the user specified
directory prefix (if any) to the set of items that determines the identity of a
given instance of a screen. Previously this was determined only by the screen
name and macro symbols/values. This identity is used when edm (e.g. for a
related display) decides to open a new instance or to raise an already open
instance (when "allow duplicates" is not selected).

x_text_obj.cc - The change on May 24, 2004 stated:

  If autosize is set, text size is updated when a display file is opened

This was not sufficient. Now, if autosize is set, text size is updated when a
display file is executed, after macro symbols values are resolved.

motifSlider.cc, motifSlider.h - Added some nice medm features. Arrow keys now
may be used to adjust the value and the control-arrow-key mutiplies the
increment by ten. The same control key multiplier works with the wheel for
wheel-mice.

main.cc - Added code to set the locale for linux where the default character set
is UTF8.