Version 1-10-0zd
color_pkg.cc, color_pkg.h, tag_pkg.cc

In previous versions, when an illegal color index is
read from a display file, it is mapped to the invalid-alarm
color. When the file is saved, the color index is likewise
modified.

Now, the read behavior is the same, but when the file is saved
the original, illegal color index is saved unchanged.

Error messages are written only to stdout. A future version will
enable errors to be displayed in a pop-up window.