Version 1-10-1h
act_win.cc - change focus behavior for overlapping objects; thanks to Larry Hoff (see his comments below)

From Larry Hoff:

John,

I discovered what I consider to be unexpected behavior in EDM under a very unusual set of circumstances. It is certainly possible that I just don't understand what the behavior should be under these circumstances. It is also true that these circumstances probably *never* happen in "real life". therefore, rather than dig deeper, and/or implement a fix, I thought I'd ask your expertise on whether this is a bug or a feature........

The issue involves detecting pointerIn/pointerOut events for overlapped widgets.

The comments in the source code for function "activeWinEventHandler" within the act_win.cc module, in the section dealing with motion events says:

"pointer in and out - processed for only the top object"

However, that seems not to be true for overlapped widgets, when the pointer moves from the "bottom" widget into the "top" widget.

The code maintains a variable called "highlightedObject", when a new widget is entered. as long as the function "enclosingObject()" remains satisfied, then "highlightedObject" will not change (i.e. as long as the pointer stays within the boundary of the widget which received the pointerIn event, it will not receive a pointerOut event). This is true even if the pointer strays into an overlapped widget which is "on top" of the previous widget.

I would expect that in the case just described, that as soon as the "upper" widget is entered, the "bottom" widget would get a pointerOut event, and the "upper" widget would get a pointerIn event. However, that would require more work in the event handler (checking for "other" enclosing objects after every mouse movement, not just checking if the highlightedObject still contains the pointer).