Version 1-10-1o
When the X Display name is read from environment, only 31 characters are
allowed. This is increased to 127.

Local PVs may be defined with a "=" or "~". The two examples below are equivalent.

LOC\one=d:1.23

LOC\one~d:1.23

This permits local pvs to be passed to a related display as macros. Instead of
sym=LOC\one=d:1.23 (which causes a problem), use sym=LOC\one~d:1.23.