Version 1-10-1q
If EDMFILTERS is set, files with foreign extensions (all except .edl, .gif, and
.png) are not opened by edm directly. Instead they are handed to a filter (if
one is registered for that extension) and the resulting stream (which must
conform to the syntax of a .edl file) is read by edm.

A file with a foreign extension may now be in the form: name[parms].ext. Before
the input filter is executed, an "is readable" check is performed. For the above
form, the check is done against name.ext. Parameters may be passed to the filter
in this manner.