Version 1-10-1zb
act_win.cc

Fixed bug introduced in 1-10-1za that causes edm to crash on reload-all
function. Thanks to David Brown (SNS).

tag_pkg.cc, button.cc

While working with Rolf Keitel's perl modules for writing edl files, I
discovered a discrepency in the widget self-generated documentation facility
with regard to default values for some properties of some widgets. This is fixed
(but more instances of this may arise).

This DOES NOT change the existing behavior of edm but will affect the
generation of edl files from utilities that use the self-generated
documentation as a guide.