Here we add the videowidget created by Larry Hoff.  The widget has also been enhanced by Diamond.  After the shared library is built one must add: TwoDProfileMonitorClass to the edmObjects file. The widget also includes some xpm files: XPixMap (XPM) consists of an ASCII image format and a C library. The format defines how to store color images (X Pixmap) in a portable and powerful way. The library provides a set of functions to store and retrieve images to and from XPM format data, being either files, buffers (files in memory), or data (included files).
While XPM is not an X Consortium standard, it is already a de facto standard. It is used in many applications both commercial and non-commercial. Several vendors distribute the XPM library, as contributed software, on the platforms they sell. Moreover, the Common Desktop Environment specifies that icons must be stored either in XBM or XPM format. Finally, Motif 2.0 from OSF includes the XPM library, allowing XPM to be used in addition to XBM.

XPM has nothing to do with dates in any form. The XPM library is therefore Y2K compliant.

This widget will be used at the SNS for the Beam Diagnostics Video foil application.