Version 1-11-1b
Makefile changes from Janet Anderson

./README.install
./calcPv/Makefile
./diamondlib/Makefile
./epicsPv/Makefile
./giflib/Makefile
./locPv/Makefile
./pnglib/Makefile

Janet's comments:

The following listing contains the few changes I made to edm version 1-11-1
to get it to build on solaris. I had to add libraries to the link step
insome of edm's Makefiles because the solaris build uses the "-z defs"
linker option by default which makes it a fatal error to have any undefined
symbols in the resulting executable or shared library. Since the solaris
shared libraries and executables have all symbols defined and the paths to
the needed libraries are recorded in resulting output file (-R or -rpath
option), LD_LIBRARY_PATH is not needed for execution of edm on solaris systems
so I modified your README slightly for solaris. I also added the same
libraries to the other os library lines in the Makefiles so it is possible
that LD_LIBRARY_PATH is not necessary for the other os systems but I did
not test it.

I cheated slightly when building because the png and gif libraries are not
installed on our solaris systems. I commented the directories out for the
solaris builds but then I redid the linux-x86 build in the png and gif
directories using the "-zdefs" linker option to see if any symbols were
undefined. If there were any, I added the resolving library to the Makefile
until there were no undefined symbols.