Version 1-12-33
expString.cc, expString.h, environment.str: added support for evaluation of symbols by a script or executable.

Example: a=1,b=file2,c=`caget -t pvname`

Subsequent references to $(c) will give the value of pvname at the time the
symbol was expanded.

Only the first line of output is read from the script and the trailing newline
is stripped.

So, for example, c=`ls *.edl` will only give a single filename when $(c) is expanded.

-----------------------

Sites that have used the "`" character for other purposes may disable this behavior by setting the environment variable EDMIGNOREEXEC to some value.

Bash example: export EDMIGNOREEXEC=1

-----------------------

I also fixed another "what was I smoking" bug in the same file. The original
intent was to ignore spaces and tabs in a symbol reference so that $(a) would be
equivalent to $( a ). This should now be the case.