Add PROXY PV type. Used to add the appearance of channel access security to channel access servers that do not implement it.
For example, given the PV named test from the portable cas, if an epics ioc is created with a PV with
the name test_PROXY, then edm will use this PV (the _PROXY one) when specified as PROXY\test in
the determination of read and write access. Unfortunately, this does not make things fundamentally
more secure and might even mislead the users into thinking otherwise.